Returns (i.e. sidechains).
And lo, upon the revision of our hoarde 5061, was the last Big Feature
committed to Three Poino, who, now more than ever, lurks imposingly on
the sidelines, heir to the throne, and eventual ruler of the realm.
His eventual succession all but guaranteed, only time and the number
of heads that must roll remain mysteries.


git-svn-id: http://subversion.ardour.org/svn/ardour2/ardour2/branches/3.0@5061 d708f5d6-7413-0410-9779-e7cbd77b26cf