More correct looking MidiSource::midi_read.
Testing seems to show the old way and this both working equally well, which is a bit weird since the old way really doesn't seem to make any sense whatsoever.
This way works as well as far as I can tell and actually makes sense, so hey, let's go with this one.


git-svn-id: http://subversion.ardour.org/svn/ardour2/ardour2/branches/3.0@5879 d708f5d6-7413-0410-9779-e7cbd77b26cf