Use evenly distributed green palette for MIDI notes. Use standard selection colour (purple) for selected MIDI notes. Green means MIDI, purple means selection, the brightness/velocity mappings match for each (just with a different hue).
Sensibly matching colour ranges brought to you by Agave, an awesome tool for this which you should use :)

I have inverted the dark/light => low/high velocity mapping for normal MIDI notes to match the selection. I don't really know which way it should be, but they should match, so I arbitrarily chose this direction.

Green also means currently selected tools and such, something we might want to resolve, but that might be at the cost of a technicolor nightmare, so perhaps not.


git-svn-id: http://subversion.ardour.org/svn/ardour2/ardour2/branches/3.0@10038 d708f5d6-7413-0410-9779-e7cbd77b26cf