Implement proper LV2 preset support.
The old code here was attempting to load several various random methods of listing LV2 presets.  IMO this situation has become untenable, the way implemented here is the way documented with examples in the extension, and if plugins don't work with it, they need to be updated.

For the record, the rationale behind this change is to make preset data separately loadable, since it tends to be far too massive to have hosts loading it for all plugins when unnecessary.

git-svn-id: http://subversion.ardour.org/svn/ardour2/ardour2/branches/3.0@11957 d708f5d6-7413-0410-9779-e7cbd77b26cf